requirements.txt
amqp==5.0.9
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
converters==0.0.1
cryptography==36.0.1
Django==2.2.5
django-cors-headers==3.10.1
django-crontab==0.7.1
django-haystack==3.1.1
django-redis==5.1.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
elasticsearch==2.4.1
fdfs-client-py==1.2.6
idna==3.3
itsdangerous==2.0.1
kombu==5.2.2
mutagen==1.40.0
Pillow==8.4.0
prompt-toolkit==3.0.24
pycparser==2.21
pycryptodomex==3.9.4
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==19.1.0
python-alipay-sdk==3.0.4
pytz==2021.3
QQLoginTool==0.3.0
redis==3.5.3
redis-py-cluster==2.1.3
requests==2.26.0
ronglian-sms-sdk==1.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
utils==1.0.1
vine==5.0.0
wcwidth==0.2.5
